move common utility functions to util.common
