+++ wilson/util/test_smeftutil.py
@@ -2,7 +2,7 @@ import unittest
 import numpy as np
 import numpy.testing as npt
 from wilson.run.smeft import beta
-from wilson.util import smeftutil
+from wilson.util import smeftutil, common
 from wilson.run.smeft.tests import test_beta
 from wilson.test_wilson import get_random_wc
 from wilson import wcxf
@@ -24,15 +24,15 @@ class TestSymm(unittest.TestCase):
 
     def test_symmetrize_symmetric(self):
         a = np.array([[1, 2, 3], [2, 4, 5], [3, 5, 6]])
-        npt.assert_array_equal(smeftutil.symmetrize_2(a), a)
+        npt.assert_array_equal(common.symmetrize_2(a), a)
         b = np.array([[1, 2, 3], [0, 4, 5], [0, 0, 6]])
-        npt.assert_array_equal(smeftutil.symmetrize_2(b), a)
+        npt.assert_array_equal(common.symmetrize_2(b), a)
 
     def test_symmetrize_hermitian(self):
         a = np.array([[1, 2j, 3j], [-2j, 4, 5j], [-3j, -5j, 6]])
-        npt.assert_array_equal(smeftutil.symmetrize_2(a), a)
+        npt.assert_array_equal(common.symmetrize_2(a), a)
         b = np.array([[1, 2j, 3j], [0, 4, 5j], [0, 0, 6]])
-        npt.assert_array_equal(smeftutil.symmetrize_2(b), a)
+        npt.assert_array_equal(common.symmetrize_2(b), a)
 
     def test_symmetrize_C(self):
         C_symm = smeftutil.symmetrize(C)
